# apps/api/public/openapi.yaml
openapi: 3.0.3
info:
  title: Guard51 API
  description: |
    Security Workforce & Operations Management Platform API.
    
    ## Authentication
    All endpoints (except /auth/login, /auth/register, /subscriptions/plans) require
    a Bearer JWT token in the Authorization header.
    
    ## Multi-tenancy
    Each request must include the tenant context via the JWT token. The API automatically
    scopes all data to the authenticated user's tenant.
    
    ## Roles
    - **super_admin**: Platform owner (DOSTHQ) — manages tenants, plans, features
    - **company_admin**: Security company administrator — full access to company data
    - **supervisor**: Site supervisor — manages guards, approves reports
    - **guard**: Security guard — clock in/out, reports, tours, chat
    - **client**: Security service client — view reports, tracking, invoices
    - **dispatcher**: Dispatch operator — manage calls, assign guards
    
  version: 1.0.0
  contact:
    name: DOSTHQ Limited
    email: api@guard51.com
    url: https://guard51.com
  license:
    name: Proprietary
    url: https://guard51.com/terms
servers:
  - url: https://api.guard51.com
    description: Production
  - url: http://localhost:8080
    description: Development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/apps/available:
    get:
      tags: [App Distribution]
      operationId: available
      summary: Available
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (insufficient role)
        '422':
          description: Validation error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
        errors:
          type: array
          items:
            type: string
...